Adição de novos filtros

+++ controllers/CardapioController.php
@@ -8,7 +8,8 @@ use app\models\CardapioSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
-
+use \yii\filters\AccessControl;
+use yii\web\ForbiddenHttpException;
 /**
  * CardapioController implements the CRUD actions for Cardapio model.
  */
@@ -17,12 +18,21 @@ class CardapioController extends Controller
     public function behaviors()
     {
         return [
-            'verbs' => [
-                'class' => VerbFilter::className(),
-                'actions' => [
-                    'delete' => ['post'],
-                ],
-            ],
+        'access' =>[
+        'class' => AccessControl::classname(),
+        'only'=> ['create','update','view','delete','index'],
+        'rules'=> [
+        ['allow'=>true,
+        'roles' => ['cardapio','index-cardapio'],
+        ],
+        ]
+        ],
+        'verbs' => [
+        'class' => VerbFilter::className(),
+        'actions' => [
+        'delete' => ['post'],
+        ],
+        ],
         ];
     }
 
@@ -32,6 +42,8 @@ class CardapioController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->can("index-cardapio") ||
+        Yii::$app->user->can("cardapio") ) {
         $searchModel = new CardapioSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -39,6 +51,9 @@ class CardapioController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
@@ -48,9 +63,14 @@ class CardapioController extends Controller
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->can("view-cardapio") ||
+            Yii::$app->user->can("cardapio") ) {
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
@@ -60,6 +80,8 @@ class CardapioController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->can("create-cardapio") ||
+            Yii::$app->user->can("cardapio") ) {
         $model = new Cardapio();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -68,6 +90,9 @@ class CardapioController extends Controller
             return $this->render('create', [
                 'model' => $model,
             ]);
+        }
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
         }
     }
 
@@ -79,6 +104,8 @@ class CardapioController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->can("update-cardapio") ||
+        Yii::$app->user->can("cardapio") ) {
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -87,6 +114,9 @@ class CardapioController extends Controller
             return $this->render('update', [
                 'model' => $model,
             ]);
+        }
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
         }
     }
 
@@ -98,9 +128,14 @@ class CardapioController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->can("delete-cardapio") ||
+        Yii::$app->user->can("cardapio") ) {
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
