+++ controllers/ComandaController.php
@@ -8,7 +8,8 @@ use app\models\ComandaSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
-
+use \yii\filters\AccessControl;
+use yii\web\ForbiddenHttpException;
 /**
  * ComandaController implements the CRUD actions for Comanda model.
  */
@@ -17,12 +18,21 @@ class ComandaController extends Controller
     public function behaviors()
     {
         return [
-            'verbs' => [
-                'class' => VerbFilter::className(),
-                'actions' => [
-                    'delete' => ['post'],
-                ],
-            ],
+        'access' =>[
+        'class' => AccessControl::classname(),
+        'only'=> ['create','update','view','delete','index'],
+        'rules'=> [
+        ['allow'=>true,
+        'roles' => ['comanda','index-comanda'],
+        ],
+        ]
+        ],
+        'verbs' => [
+        'class' => VerbFilter::className(),
+        'actions' => [
+        'delete' => ['post'],
+        ],
+        ],
         ];
     }
 
@@ -32,6 +42,9 @@ class ComandaController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->can("index-comanda") ||
+        Yii::$app->user->can("comanda") ) {
+        
         $searchModel = new ComandaSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -39,6 +52,9 @@ class ComandaController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
@@ -48,9 +64,15 @@ class ComandaController extends Controller
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->can("view-comanda") ||
+            Yii::$app->user->can("comanda") ) {
+        
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
@@ -60,6 +82,9 @@ class ComandaController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->can("create-comanda") ||
+            Yii::$app->user->can("comanda") ) {
+        
         $model = new Comanda();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -68,6 +93,9 @@ class ComandaController extends Controller
             return $this->render('create', [
                 'model' => $model,
             ]);
+        }
+        } else{
+            throw new ForbiddenHttpException("Acesso negado!");
         }
     }
 
@@ -79,6 +107,9 @@ class ComandaController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->can("update-comanda") ||
+        Yii::$app->user->can("comanda") ) {
+        
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -87,6 +118,9 @@ class ComandaController extends Controller
             return $this->render('update', [
                 'model' => $model,
             ]);
+        }
+        } else{
+            throw new ForbiddenHttpException("Acesso negado!");
         }
     }
 
@@ -98,9 +132,15 @@ class ComandaController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->can("delete-comanda") ||
+        Yii::$app->user->can("comanda") ) {
+        
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        } else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
