+++ controllers/HistoricosituacaoController.php
@@ -17,13 +17,23 @@ class HistoricosituacaoController extends Controller
     public function behaviors()
     {
         return [
-            'verbs' => [
-                'class' => VerbFilter::className(),
-                'actions' => [
-                    'delete' => ['post'],
-                ],
-            ],
+        'access' =>[
+        'class' => AccessControl::classname(),
+        'only'=> ['create','update','view','delete','index'],
+        'rules'=> [
+        ['allow'=>true,
+        'roles' => ['historicosituacao','index-historicosituacao'],
+        ],
+        ]
+        ],
+        'verbs' => [
+        'class' => VerbFilter::className(),
+        'actions' => [
+        'delete' => ['post'],
+        ],
+        ],
         ];
+
     }
 
     /**
@@ -32,6 +42,9 @@ class HistoricosituacaoController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->can("index-historicosituacao") ||
+        Yii::$app->user->can("historicosituacao") ) {
+
         $searchModel = new HistoricosituacaoSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -39,6 +52,9 @@ class HistoricosituacaoController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
@@ -49,9 +65,15 @@ class HistoricosituacaoController extends Controller
      */
     public function actionView($idPedido, $idSituacaoPedido)
     {
+        if (Yii::$app->user->can("view-historicosituacao") ||
+        Yii::$app->user->can("historicosituacao") ) {
+
         return $this->render('view', [
             'model' => $this->findModel($idPedido, $idSituacaoPedido),
         ]);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
@@ -61,6 +83,9 @@ class HistoricosituacaoController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->can("create-historicosituacao") ||
+        Yii::$app->user->can("historicosituacao") ) {
+
         $model = new Historicosituacao();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -69,6 +94,9 @@ class HistoricosituacaoController extends Controller
             return $this->render('create', [
                 'model' => $model,
             ]);
+        }
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
         }
     }
 
@@ -81,6 +109,9 @@ class HistoricosituacaoController extends Controller
      */
     public function actionUpdate($idPedido, $idSituacaoPedido)
     {
+        if (Yii::$app->user->can("update-historicosituacao") ||
+        Yii::$app->user->can("historicosituacao") ) {
+
         $model = $this->findModel($idPedido, $idSituacaoPedido);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -89,6 +120,9 @@ class HistoricosituacaoController extends Controller
             return $this->render('update', [
                 'model' => $model,
             ]);
+        }
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
         }
     }
 
@@ -101,9 +135,15 @@ class HistoricosituacaoController extends Controller
      */
     public function actionDelete($idPedido, $idSituacaoPedido)
     {
+        if (Yii::$app->user->can("delete-historicosituacao") ||
+        Yii::$app->user->can("historicosituacao") ) {
+
         $this->findModel($idPedido, $idSituacaoPedido)->delete();
 
         return $this->redirect(['index']);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
