+++ controllers/ItemcardapioController.php
@@ -8,6 +8,8 @@ use app\models\ItemcardapioSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\NotFoundHttpException;
+use yii\filters\VerbFilter;
 
 /**
  * ItemcardapioController implements the CRUD actions for Itemcardapio model.
@@ -17,12 +19,21 @@ class ItemcardapioController extends Controller
     public function behaviors()
     {
         return [
-            'verbs' => [
-                'class' => VerbFilter::className(),
-                'actions' => [
-                    'delete' => ['post'],
-                ],
-            ],
+        'access' =>[
+        'class' => AccessControl::classname(),
+        'only'=> ['create','update','view','delete','index'],
+        'rules'=> [
+        ['allow'=>true,
+        'roles' => ['itemcardapio','index-itemcardapio'],
+        ],
+        ]
+        ],
+        'verbs' => [
+        'class' => VerbFilter::className(),
+        'actions' => [
+        'delete' => ['post'],
+        ],
+        ],
         ];
     }
 
@@ -32,6 +43,9 @@ class ItemcardapioController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->can("index-itemcardapio") ||
+        Yii::$app->user->can("itemcardapio") ) {
+
         $searchModel = new ItemcardapioSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -39,6 +53,9 @@ class ItemcardapioController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
@@ -49,9 +66,15 @@ class ItemcardapioController extends Controller
      */
     public function actionView($idCardapio, $idProduto)
     {
+        if (Yii::$app->user->can("view-itemcardapio") ||
+        Yii::$app->user->can("itemcardapio") ) {
+
         return $this->render('view', [
             'model' => $this->findModel($idCardapio, $idProduto),
         ]);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
@@ -61,6 +84,9 @@ class ItemcardapioController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->can("create-itemcardapio") ||
+        Yii::$app->user->can("itemcardapio") ) {
+
         $model = new Itemcardapio();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -69,6 +95,9 @@ class ItemcardapioController extends Controller
             return $this->render('create', [
                 'model' => $model,
             ]);
+        }
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
         }
     }
 
@@ -81,6 +110,9 @@ class ItemcardapioController extends Controller
      */
     public function actionUpdate($idCardapio, $idProduto)
     {
+        if (Yii::$app->user->can("update-itemcardapio") ||
+        Yii::$app->user->can("itemcardapio") ) {
+
         $model = $this->findModel($idCardapio, $idProduto);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -89,6 +121,9 @@ class ItemcardapioController extends Controller
             return $this->render('update', [
                 'model' => $model,
             ]);
+        }
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
         }
     }
 
@@ -101,9 +136,15 @@ class ItemcardapioController extends Controller
      */
     public function actionDelete($idCardapio, $idProduto)
     {
+        if (Yii::$app->user->can("delete-itemcardapio") ||
+        Yii::$app->user->can("itemcardapio") ) {
+
         $this->findModel($idCardapio, $idProduto)->delete();
 
         return $this->redirect(['index']);
+        }else{
+            throw new ForbiddenHttpException("Acesso negado!");
+        }
     }
 
     /**
